chore: onInstall hook work (#145)
In order for users to have "by default" a Solana account we are calling
for its creation when the snap gets installed.
This PR adds the snap
[onInstall](https://docs.metamask.io/snaps/reference/entry-points/#oninstall)
hook in order to automatically create a Solana account when:
1. There isn't any account yet in keyring, we create a new one from
scratch.
2. The user installs MM and imports an existing SRP and already has a
Solana account associated to that seed phrase.
For this case we will search 5 times the derivation path for an address
with a SOL balance, and if we find it, we add that account.

This hook triggers when the snap gets installed in the extension.

As a follow up work, after the multi-SRP initiative, I think we will
also need to trigger this workflow every time the user imports a new
SRP.