fix: restore max button in send flow (#384)
- Rework the fee management for Send flows by removing the RPC call to
`getComputeUnitEstimate`, and using hard coded compute unit limits.
Since we are in control of the tx instructions, the compute units are
predictable.
- Restore the `Max` button in Send form, and fix the max calculation:
- Subtract 1 extra lamport to the max, to account for precision loss
when computing priority fees. This ensures we never end up trying to
send more than possible.
- Improve validation for sending SPL token. The amount is now invalid if
the user's balance would drop below the rent after paying for the tx
fee.
- Improve module encapsulation by moving the `SendSolBuild` and
`SendSplTokenBuilder` inside the `features/send`, since this is the only
place where they will be used.
- Improve the builders interface to leverage polymorphism.
- Fix an issue in `AssetsService` where saving new balances to state
would override the existing ones.